♻️ refactor(nav-panel): move the shell and the Home panel export into features
`src/routes/` may only hold thin route segments, so `NavPanelShell` did not
belong there — and it reached directly into another route's private
`home/_layout/Sidebar`.

Move the shell to `@/features/NavPanel/Shell` and give the Home panel a feature
module of its own (`@/features/HomeSidebar`), mirroring `AgentSidebar`. Both main
layout twins now compose that feature API and nothing else.
