+++ src/routes/(main)/_layout/index.tsx
@@ -24,6 +24,7 @@ import ZoomHUD from '@/features/Electron/system/ZoomHUD';
 import TabCacheBridges from '@/features/Electron/titlebar/TabBar/TabCacheBridges';
 import TitleBar from '@/features/Electron/titlebar/TitleBar';
 import HotkeyHelperPanel from '@/features/HotkeyHelperPanel';
+import NavPanelShell from '@/features/NavPanel/Shell';
 import { RouteMetaBridge } from '@/features/RouteMeta';
 import { usePlatform } from '@/hooks/usePlatform';
 import CmdkLazy from '@/layout/GlobalProvider/CmdkLazy';
@@ -35,7 +36,6 @@ import DesktopHome from '../home';
 import DesktopHomeLayout from '../home/_layout';
 import DesktopAutoOidcOnFirstOpen from './DesktopAutoOidcOnFirstOpen';
 import DesktopLayoutContainer from './DesktopLayoutContainer';
-import NavPanelShell from './NavPanelShell';
 import RegisterHotkeys from './RegisterHotkeys';
 import { styles } from './style';
