--- src/routes/(main)/home/_layout/SidebarContent.tsx
@@ -1,17 +1,0 @@
-import { memo } from 'react';
-
-import SideBarLayout from '@/features/NavPanel/SideBarLayout';
-
-import Body from './Body';
-import { AgentModalProvider } from './Body/Agent/ModalProvider';
-import Header from './Header';
-
-const Sidebar = memo(() => {
-  return (
-    <AgentModalProvider>
-      <SideBarLayout body={<Body />} header={<Header />} />
-    </AgentModalProvider>
-  );
-});
-
-export default Sidebar;
